Microservice implementation + code linting and cleaning
