+++ package.json
@@ -1,18 +1,26 @@
 {
-    "name": "server",
+    "name": "chain-bridge",
     "version": "0.1.0",
-    "description": "",
+    "description": "Glorious gateway to the Lum Network blockchain",
     "author": "",
     "private": true,
-    "license": "UNLICENSED",
+    "license": "apache",
     "scripts": {
         "prebuild": "rimraf dist",
-        "build": "nest build",
+        "build": "nest build api",
+        "build:api": "nest build api",
+        "build:sync:consumer": "nest build sync:consumer",
+        "build:sync:scheduler": "nest build sync:scheduler",
         "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-        "start": "nest start",
-        "start:dev": "nest start --watch",
-        "start:debug": "nest start --debug --watch",
-        "start:prod": "node dist/main",
+        "start:api:dev": "nest start api --watch",
+        "start:api:debug": "nest start api --debug --watch",
+        "start:api:prod": "node dist/api",
+        "start:sync:consumer:dev": "nest start sync:consumer --watch",
+        "start:sync:consumer:debug": "nest start sync:consumer --debug --watch",
+        "start:sync:consumer:prod": "node dist/sync_scheduler",
+        "start:sync:scheduler:dev": "nest start sync:scheduler --watch",
+        "start:sync:scheduler:debug": "nest start sync:scheduler --debug --watch",
+        "start:sync:scheduler:prod": "node dist/sync_scheduler",
         "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
         "test": "jest",
         "test:watch": "jest --watch",
@@ -23,15 +31,16 @@
     "dependencies": {
         "@elastic/elasticsearch": "^7.9.1",
         "@lum-network/sdk-javascript": "^0.7.5",
-        "@nestjs/axios": "^0.0.3",
-        "@nestjs/bull": "^0.2.2",
-        "@nestjs/common": "^7.0.0",
-        "@nestjs/core": "^7.0.0",
-        "@nestjs/platform-express": "^7.0.0",
-        "@nestjs/platform-socket.io": "^7.6.13",
-        "@nestjs/schedule": "^0.4.1",
-        "@nestjs/terminus": "^7.0.1",
-        "@nestjs/websockets": "^7.6.13",
+        "@nestjs/axios": "^0.0.7",
+        "@nestjs/bull": "^0.5.2",
+        "@nestjs/common": "^8.4.0",
+        "@nestjs/core": "^8.4.0",
+        "@nestjs/microservices": "^8.4.0",
+        "@nestjs/platform-express": "^8.4.0",
+        "@nestjs/platform-socket.io": "^8.4.0",
+        "@nestjs/schedule": "^1.0.2",
+        "@nestjs/terminus": "^8.0.4",
+        "@nestjs/websockets": "^8.4.0",
         "bull": "^3.18.0",
         "cache-manager": "^3.4.0",
         "cache-manager-redis-store": "^2.0.0",
@@ -44,33 +53,33 @@
         "node-notifier": "^8.0.1",
         "reflect-metadata": "^0.1.13",
         "rimraf": "^3.0.2",
-        "rxjs": "^6.5.4",
+        "rxjs": "^7.5.5",
         "xmlhttprequest-ssl": "^1.6.2"
     },
     "devDependencies": {
-        "@nestjs/cli": "^7.0.0",
-        "@nestjs/schematics": "^7.0.0",
-        "@nestjs/testing": "^7.0.0",
-        "@types/bull": "^3.14.4",
-        "@types/express": "^4.17.3",
+        "@nestjs/cli": "^8.2.2",
+        "@nestjs/schematics": "^8.0.8",
+        "@nestjs/testing": "^8.4.0",
+        "@types/bull": "^3.15.8",
+        "@types/express": "^4.17.13",
         "@types/jest": "25.1.4",
         "@types/moment": "^2.13.0",
-        "@types/node": "^13.9.1",
-        "@types/socket.io": "^2.1.13",
-        "@types/supertest": "^2.0.8",
+        "@types/node": "^17.0.21",
+        "@types/socket.io": "^3.0.2",
+        "@types/supertest": "^2.0.11",
         "@typescript-eslint/eslint-plugin": "^2.23.0",
         "@typescript-eslint/parser": "^2.23.0",
         "eslint": "^6.8.0",
         "eslint-config-prettier": "^6.10.0",
         "eslint-plugin-import": "^2.20.1",
         "jest": "^25.1.0",
-        "prettier": "^1.19.1",
-        "supertest": "^4.0.2",
+        "prettier": "^2.5.1",
+        "supertest": "^6.2.2",
         "ts-jest": "25.2.1",
         "ts-loader": "^6.2.1",
-        "ts-node": "^8.6.2",
+        "ts-node": "^10.7.0",
         "tsconfig-paths": "^3.9.0",
-        "typescript": "^3.7.4"
+        "typescript": "^4.6.2"
     },
     "jest": {
         "moduleFileExtensions": [
