+++ src/http/interceptors/response.interceptor.ts
@@ -12,7 +12,7 @@ export interface Response<T> {
 export class ResponseInterceptor<T> implements NestInterceptor<T, Response<T>> {
     intercept(context: ExecutionContext, next: CallHandler): Observable<Response<T>> {
         return next.handle().pipe(
-            map(data => ({
+            map((data) => ({
                 code: 200,
                 message: '',
                 result: data,
