+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './elastic.service';
 export * from './lum-network.service';
+export * from './lum.service';
