+++ src/services/lum-network.service.ts
@@ -1,4 +1,4 @@
-import {Injectable, Logger} from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { LumClient } from '@lum-network/sdk-javascript';
 
 import { config } from '@app/utils/config';
@@ -12,7 +12,7 @@ export class LumNetworkService {
         try {
             this._client = await LumClient.connect(config.getLumNetworkEndpoint());
             const chainId = await this._client.getChainId();
-            this._logger.log(`Connection established to Lum Network on ${config.getLumNetworkEndpoint()} = ${chainId}`)
+            this._logger.log(`Connection established to Lum Network on ${config.getLumNetworkEndpoint()} = ${chainId}`);
         } catch (e) {
             console.error(e);
         }
