+++ test/app.e2e-spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { AppModule } from './../src/app.module';
+import { ApiModule } from '@app/modules/api.module';
 
 describe('AppController (e2e)', () => {
     let app: INestApplication;
 
     beforeEach(async () => {
         const moduleFixture: TestingModule = await Test.createTestingModule({
-            imports: [AppModule],
+            imports: [ApiModule],
         }).compile();
 
         app = moduleFixture.createNestApplication();
@@ -16,9 +16,6 @@ describe('AppController (e2e)', () => {
     });
 
     it('/ (GET)', () => {
-        return request(app.getHttpServer())
-            .get('/')
-            .expect(200)
-            .expect('Hello World!');
+        return request(app.getHttpServer()).get('/').expect(200).expect('Hello World!');
     });
 });
